+++ web/app/views/license_mailer/imageflow_id_license_email.text.erb
