+++ web/app/views/license_mailer/imageresizer_id_license_email.text.erb
