--- web/old_rails/license_builder.rb
@@ -1,9 +1,0 @@
-class LicenseBuilder
-
-  attr_accessor :parse
-  def initialize(parse)
-    self.parse = parse
-  end
-
-  
-end
